refactor: extracted commaon osv-scalibr's function in scalibr.go
